+++ lib/pl-zip-codes.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "pl_zip_codes"