+++ lib/pl_zip_codes/errors.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module PlZipCodes
-  class Error < StandardError; end
-
-  class DownloadError < Error; end
-
-  class DatasetError < Error; end
-end