+++ lib/zip_codes/pl/errors.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module ZipCodes
+  module PL
+    class Error < StandardError; end
+
+    class DownloadError < Error; end
+
+    class DatasetError < Error; end
+  end
+end